requirements.txt
flask[async]
requests
pact-python
pytest
yarl
python-dotenv